Project/Back-end_Flask/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
pytest==8.0.2
requests==2.31.0 